Move Faucet class where it belongs
